refactor(pico8): move the player helpers into v2/utils
The PICO-8 modules sat beside the view they serve, while every other v2
helper lives in v2/utils, player ones included (playerBezel, playerDisc,
playerFirmware).

scriptLoader moves with them: pico8Runtime imports it, and nothing in
v2/utils reaches into v2/views, so leaving it behind would have been the
first exception. EmulatorJS and JsDos pick it up from the new path.
