+++ frontend/src/v2/utils/scriptLoader.ts
