+++ frontend/src/v2/views/Player/EmulatorJS.vue
@@ -71,9 +71,9 @@ import {
   type DiscSelection,
 } from "@/v2/utils/playerDisc";
 import { resolveInitialFirmware } from "@/v2/utils/playerFirmware";
+import { isJsResource, loadScript } from "@/v2/utils/scriptLoader";
 import { installIOSFullscreenShim } from "@/views/Player/EmulatorJS/utils";
 import { rememberCore, resolveRememberedCore } from "./coreStorage";
-import { isJsResource, loadScript } from "./scriptLoader";
 
 // Reuse v1's heavy emulator integration — do NOT rewrite this. Lazy so the
 // bundle doesn't pull in the EJS shims until we actually mount the player.
